api/requirements.txt
adal==1.2.7
async-timeout==4.0.3
azure-common==1.1.28
azure-core==1.29.4
azure-keyvault==1.1.0
azure-keyvault-secrets==4.7.0
azure-nspkg==3.0.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.4
idna==3.4
isodate==0.6.1
msrest==0.7.1
msrestazure==0.6.4
oauthlib==3.2.2
praw==7.7.1
prawcore==2.3.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
spotipy==2.23.0
typing_extensions==4.8.0
update-checker==0.18.0
urllib3==2.0.5
websocket-client==1.6.3
azure-functions
azure-cosmos